+++ backpack/core/derivatives/tanh.py
@@ -20,7 +20,7 @@ class TanhDerivatives(ElementwiseDerivatives):
         subsampling: List[int] = None,
     ) -> Tensor:
         output = subsample(module.output, subsampling=subsampling)
-        return 1.0 - output ** 2
+        return 1.0 - output**2
 
     def d2f(self, module, g_inp, g_out):
-        return -2.0 * module.output * (1.0 - module.output ** 2)
+        return -2.0 * module.output * (1.0 - module.output**2)