+++ backpack/extensions/secondorder/diag_ggn/__init__.py
@@ -48,8 +48,10 @@ from torch.nn import (
 )
 
 from backpack.custom_module.branching import SumModule
+from backpack.custom_module.pad import Pad
 from backpack.custom_module.permute import Permute
 from backpack.custom_module.scale_module import ScaleModule
+from backpack.custom_module.slicing import Slicing
 from backpack.extensions.secondorder.base import SecondOrderBackpropExtension
 from backpack.extensions.secondorder.hbp import LossHessianStrategy
 
@@ -69,10 +71,12 @@ from . import (
     flatten,
     linear,
     losses,
+    pad,
     padding,
     permute,
     pooling,
     rnn,
+    slicing,
 )
 
 
@@ -143,6 +147,8 @@ class DiagGGN(SecondOrderBackpropExtension):
                 BatchNorm2d: batchnorm_nd.DiagGGNBatchNormNd(),
                 BatchNorm3d: batchnorm_nd.DiagGGNBatchNormNd(),
                 Embedding: embedding.DiagGGNEmbedding(),
+                Pad: pad.DiagGGNPad(),
+                Slicing: slicing.DiagGGNSlicing(),
             },
         )
 
@@ -266,6 +272,8 @@ class BatchDiagGGN(SecondOrderBackpropExtension):
                 BatchNorm2d: batchnorm_nd.BatchDiagGGNBatchNormNd(),
                 BatchNorm3d: batchnorm_nd.BatchDiagGGNBatchNormNd(),
                 Embedding: embedding.BatchDiagGGNEmbedding(),
+                Pad: pad.DiagGGNPad(),
+                Slicing: slicing.DiagGGNSlicing(),
             },
         )